requirements.txt
rpy2
click
pandas
ibis-framework[sqlite,duckdb]
sqlparse
python-dateutil
duckdb
duckdb-engine
